+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Move the `alchemy-select` pagination checker initializer to a static function
 * Prevent `alchemy-select` from keeping on loading remote data when last few results were empty
+* Fix pagination breaking re-opening of `alchemy-select` dropdown
 
 ## 0.1.7 (2022-06-12)